backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pdfplumber==0.10.3
python-multipart==0.0.6
sqlalchemy==2.0.36
google-generativeai==0.3.1
python-dotenv==1.0.0
python-docx==1.1.0
pypandoc==1.12
alembic==1.13.2